shopping-cart/requirements.txt
flask[async]
pymongo[srv]
python-dateutil
pytest-asyncio
flask_cors